requirements.txt
fastapi
uvicorn[standard]   
python-multipart
SQLAlchemy
python-jose
pydantic[email]
passlib[bcrypt]
pytz